First cut porting pvlib_detect_shadows from matlab
This is a rought implementation that returns almost the correct shadow
image. There is still some work to be done. The output needs to be
returned as a series instead of an image and steps should be taken to
minimize memory use, especially in the _clean_wires() function, by
passing ndarrays rather than allocating a new array for every step.

Currently this implementation appears to be more sensitive than the
matlab version, and includes more remnants of clouds in the output.