Filter blobs by height (e.g. days) rather than width
This is consistent with the comments in the original matlab
code, although it is not 100% clear to me that it matches the
implementation. Since it produces results that are closer to
the output (in addition to matching the comments) from the MATLAB
code, I will assume that my reading of the original
implementation was not correct.