Specify output array for image operations in _filter_bars()
Without specifying an output array the image processing and numpy
functions will allocate a new array for every operation. This can
be a very expensive operation; fortunately we can avoid it by using
a temporary array for the morphological operations and specifying
an output array for the resulting image with horizontal bars
removed.