Detect shadows from fixed objects in GHI data (#101)
* First cut porting pvlib_detect_shadows from matlab

This is a rought implementation that returns almost the correct shadow
image. There is still some work to be done. The output needs to be
returned as a series instead of an image and steps should be taken to
minimize memory use, especially in the _clean_wires() function, by
passing ndarrays rather than allocating a new array for every step.

Currently this implementation appears to be more sensitive than the
matlab version, and includes more remnants of clouds in the output.

* Add test module for features.shading

Initially adds basic tests for data without any fixed shadows.

* Correct order of masks passed to binary hit/miss filter

In the remove_pillars step the masks were passed in the wrong
order.

* Filter blobs by height (e.g. days) rather than width

This is consistent with the comments in the original matlab
code, although it is not 100% clear to me that it matches the
implementation. Since it produces results that are closer to
the output (in addition to matching the comments) from the MATLAB
code, I will assume that my reading of the original
implementation was not correct.

* Clean up variable naming and comments in shading.fixed()

* Tweak shading._filter_bars()

Changes to better match the original implementation. Still not
completely sure that this correct though.

* Apply _filter_bars() to the raw shadow image

Rather than applying it on successively more cleaned up images.

* Make min_gradient a parameter to shading.fixed()

* Use morphology.remove_small_objects in initial shadow detection

This should be slightly more efficient, plus it matched the intent
of the code better.

Also updates name of structuring element to match its shape

* Make the minimum duration of cloudy times a parameter

* Return shadows as a series and an image

Converts the image back to a series and re-indexes it like
the input series.

* Fix errors in test_shading.test_fixed_same_index

* Test that a fake shadow at the same time every day is detected

* Fix grammar in comment

* Add scikit-image to dependencies

* Use a 3x3 mask in shading._smooth()

* Remove extra calls to _to_image() in shading._prepare_images()

* Docstring and comment updates

* Raise a ValueError for data with invalid intervals

* Apply docstring edits from code review

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* Rename `window` to `duration` in shading._detect_clouds()

Adds a better docstring for the returned series.

* Bump minimum pandas version to 0.24.0

Following NEP 29, we can now require at least 0.24.0 which was
released 24 months ago.

* Add parameter descriptions to shading._prepare_images()

* Specify output array for image operations in _filter_bars()

Without specifying an output array the image processing and numpy
functions will allocate a new array for every operation. This can
be a very expensive operation; fortunately we can avoid it by using
a temporary array for the morphological operations and specifying
an output array for the resulting image with horizontal bars
removed.

* Remove unused shading._width() function

* Pass output array to _fill_gaps() and _remove_spikes()

Improves performance by reducing the number of new arrays that must
be allocated.

* debugging

Transpose the masks used in `_detect_pillars`. The original matlab image was transpose
of what we are working with here.

* Tidy up doc string

Add image to return values

* Add shading to api.rst and whatsnew

* Docstring edits

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* Add features.shading to package overview in index.rst

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>